Sprint 14 feat : add entity for User and Item

+++ src/main/java/ru/practicum/shareit/item/dto/ItemDto.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 @Builder
 public class ItemDto {
-    private int id;
-    private int userId;
+    private Long id;
+    private Long userId;
     private String name;
     private String description;
     private Boolean available;
